+++ packages/translations/src/translations/index.ts
@@ -3,9 +3,11 @@ import { enTranslations } from './en';
 import { esTranslations } from './es';
 import { fiTranslations } from './fi';
 import { frTranslations } from './fr';
+import { idTranslations } from './id';
 import { itTranslations } from './it';
 import { nlTranslations } from './nl';
 import { ptTranslations } from './pt';
+import { zhTranslations } from './zh';
 
 const baseTranslations = {
 	en: enTranslations,
@@ -13,9 +15,11 @@ const baseTranslations = {
 	es: esTranslations,
 	fi: fiTranslations,
 	fr: frTranslations,
+	id: idTranslations,
 	it: itTranslations,
 	nl: nlTranslations,
 	pt: ptTranslations,
+	zh: zhTranslations,
 };
 
 export { baseTranslations };